chore: Refactor ApplicationView.SetVisibleBounds() to shared code
